+++ core/providers/anthropic/responses.go
@@ -1419,12 +1419,18 @@ func ToAnthropicResponsesRequest(bifrostReq *schemas.BifrostResponsesRequest) (*
 		}
 		if bifrostReq.Params.Reasoning != nil {
 			if bifrostReq.Params.Reasoning.MaxTokens != nil {
-				if *bifrostReq.Params.Reasoning.MaxTokens < MinimumReasoningMaxTokens {
+				budgetTokens := *bifrostReq.Params.Reasoning.MaxTokens
+				if *bifrostReq.Params.Reasoning.MaxTokens == -1 {
+					// anthropic does not support dynamic reasoning budget like gemini
+					// setting it to default max tokens
+					budgetTokens = MinimumReasoningMaxTokens
+				}
+				if budgetTokens < MinimumReasoningMaxTokens {
 					return nil, fmt.Errorf("reasoning.max_tokens must be >= %d for anthropic", MinimumReasoningMaxTokens)
 				}
 				anthropicReq.Thinking = &AnthropicThinking{
 					Type:         "enabled",
-					BudgetTokens: bifrostReq.Params.Reasoning.MaxTokens,
+					BudgetTokens: schemas.Ptr(budgetTokens),
 				}
 			} else {
 				if bifrostReq.Params.Reasoning.Effort != nil {