+++ core/providers/bedrock/responses.go
@@ -1468,13 +1468,19 @@ func ToBedrockResponsesRequest(ctx *context.Context, bifrostReq *schemas.Bifrost
 				bedrockReq.AdditionalModelRequestFields = make(schemas.OrderedMap)
 			}
 			if bifrostReq.Params.Reasoning.MaxTokens != nil {
-				if schemas.IsAnthropicModel(bifrostReq.Model) && *bifrostReq.Params.Reasoning.MaxTokens < anthropic.MinimumReasoningMaxTokens {
+				tokenBudget := *bifrostReq.Params.Reasoning.MaxTokens
+				if *bifrostReq.Params.Reasoning.MaxTokens == -1 {
+					// bedrock does not support dynamic reasoning budget like gemini
+					// setting it to default max tokens
+					tokenBudget = anthropic.MinimumReasoningMaxTokens
+				}
+				if schemas.IsAnthropicModel(bifrostReq.Model) && tokenBudget < anthropic.MinimumReasoningMaxTokens {
 					return nil, fmt.Errorf("reasoning.max_tokens must be >= %d for anthropic", anthropic.MinimumReasoningMaxTokens)
 				}
 				if schemas.IsAnthropicModel(bifrostReq.Model) {
 					bedrockReq.AdditionalModelRequestFields["reasoning_config"] = map[string]any{
 						"type":          "enabled",
-						"budget_tokens": *bifrostReq.Params.Reasoning.MaxTokens,
+						"budget_tokens": tokenBudget,
 					}
 				} else if schemas.IsNovaModel(bifrostReq.Model) {
 					minBudgetTokens := MinimumReasoningMaxTokens
@@ -1484,7 +1490,7 @@ func ToBedrockResponsesRequest(ctx *context.Context, bifrostReq *schemas.Bifrost
 					} else {
 						inferenceConfig.MaxTokens = schemas.Ptr(DefaultCompletionMaxTokens)
 					}
-					maxReasoningEffort := providerUtils.GetReasoningEffortFromBudgetTokens(*bifrostReq.Params.Reasoning.MaxTokens, minBudgetTokens, defaultMaxTokens)
+					maxReasoningEffort := providerUtils.GetReasoningEffortFromBudgetTokens(tokenBudget, minBudgetTokens, defaultMaxTokens)
 					typeStr := "enabled"
 					switch maxReasoningEffort {
 					case "high":